Convert `<FilteredNotificationsBanner>` to Typescript
